CodeFactor: Move argument completer to another file
