+++ src/Microsoft.PowerShell.Commands.Utility/commands/utility/WebCmdlet/Common/WebRequestPSCmdlet.Common.cs
@@ -8,7 +8,6 @@ using System.Collections.ObjectModel;
 using System.Globalization;
 using System.IO;
 using System.Management.Automation;
-using System.Management.Automation.Language;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
@@ -24,8 +23,6 @@ using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Xml;
-using System.Net.Mime;
-using System.Reflection;
 
 namespace Microsoft.PowerShell.Commands
 {
@@ -1910,58 +1907,6 @@ namespace Microsoft.PowerShell.Commands
         };
 
         #endregion Helper Methods
-    }
-
-    /// <summary>
-    /// Provides argument completion for ContentType parameter.
-    /// </summary>
-    public class ContentTypeArgumentCompleter : IArgumentCompleter
-    {
-        /// <summary>
-        /// Returns completion results for ContentType parameter.
-        /// </summary>
-        /// <param name="commandName">The command name.</param>
-        /// <param name="parameterName">The parameter name.</param>
-        /// <param name="wordToComplete">The word to complete.</param>
-        /// <param name="commandAst">The command AST.</param>
-        /// <param name="fakeBoundParameters">The fake bound parameters.</param>
-        /// <returns>List of Completion Results.</returns>
-        public IEnumerable<CompletionResult> CompleteArgument(
-            string commandName,
-            string parameterName,
-            string wordToComplete,
-            CommandAst commandAst,
-            IDictionary fakeBoundParameters)
-                => CompletionCompleters.GetMatchingResults(
-                    wordToComplete,
-                    possibleCompletionValues: EnumerateMediaTypeFieldValues(
-                        typeof(MediaTypeNames.Application),
-                        typeof(MediaTypeNames.Text)),
-                    resultType: CompletionResultType.ParameterValue);
-
-        /// <summary>
-        /// Enumerate media type field values.
-        /// This uses reflection to extract the public field string values from an array of media types.
-        /// </summary>
-        /// <param name="mediaTypes">The array of media types.</param>
-        /// <returns>Enumerator of media type field value strings.</returns>
-        private static IEnumerable<string> EnumerateMediaTypeFieldValues(params Type[] mediaTypes)
-        {
-            foreach (Type mediaType in mediaTypes)
-            {
-                foreach (FieldInfo field in mediaType.GetFields(BindingFlags.Public | BindingFlags.Static))
-                {
-                    if (field.FieldType == typeof(string))
-                    {
-                        string fieldValue = (string)field.GetValue(null);
-                        if (!string.IsNullOrEmpty(fieldValue))
-                        {
-                            yield return fieldValue;
-                        }
-                    }
-                }
-            }
-        }
     }
 
     /// <summary>
